Moved common code for migration examples to separate module, removed redundant methods
